--- src/routes/components/Header.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-	import HeaderDesktop from './headerComponents/HeaderDesktop.svelte';
-	import HeaderMobile from './headerComponents/HeaderMobile.svelte';
-</script>
-
-<HeaderDesktop />
-<HeaderMobile />
