+++ src/routes/componentsHomePage/Header.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+	import HeaderDesktop from './componentsHeader/HeaderDesktop.svelte';
+	import HeaderMobile from './componentsHeader/HeaderMobile.svelte';
+</script>
+
+<HeaderDesktop />
+<HeaderMobile />
