+++ src/routes/componentsHomePage/componentSettings/i18n.ts
