+++ src/routes/componentsHomePage/componentsHeader/HeaderMobile.svelte
