+++ src/routes/componentsHomePage/i18n.ts
