+++ src/routes/index.svelte
@@ -1,5 +1,5 @@
 <script>
-	import Content from './components/Content.svelte';
+	import Content from './componentsHomePage/Content.svelte';
 </script>
 
 <Content>
